GMsFEM works
Fixed bugs (offset_x was not turned zero in one loop) and major error -
incorrect indexation of the R matrix (the indexation of the global R matrix has
to correspond to the global indexation - so that it can be multiplied with
global fine scale matrices and vectors). All these bugs were fixed, and it
seems that GMsFEM works now. It has to be tested, though.